requirements.txt
asgiref==3.7.2
awsebcli==3.20.7
botocore==1.29.158
cachetools==5.3.1
cement==2.8.2
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.3
distlib==0.3.6
Django==2.2
django-bootstrap-modal-forms==2.2.1
django-crispy-forms==1.14.0
django-environ==0.10.0
filelock==3.12.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
libcst==1.0.0
MarkupSafe==2.1.3
mpmath==1.3.0
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
pathspec==0.10.1
Pillow==9.5.0
platformdirs==3.8.0
proto-plus==1.22.2
protobuf==4.23.2
psycopg2==2.9.6
pypiwin32==223
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==5.3.1
requests==2.31.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.4
sympy==1.12
termcolor==1.1.0
torch==2.0.1
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.23.1
wcwidth==0.1.9
